Правки по проверке типов и убрал лишний dispatch
