+++ src/components/modal-overlay/modal-overlay.jsx
@@ -1,22 +1,9 @@
 import overlaySt from "./overlay.module.css";
 import PropTypes from "prop-types";
-import React, { useEffect } from "react";
-const ESC_KEYCODE = 27;
 
 function ModalOverlay(props) {
   const { children, onClose } = props;
 
-  useEffect(() => {
-    const closeModal = (e) => {
-      if (e.key === "Escape" || e.keyCode === ESC_KEYCODE) {
-        onClose();
-      }
-    };
-    document.addEventListener("keydown", closeModal);
-    return () => {
-      document.removeEventListener("keydown", closeModal);
-    };
-  }, [onClose]);
   return (
     <div className={overlaySt.modal} onClick={() => onClose()}>
       {children}
@@ -25,5 +12,6 @@ function ModalOverlay(props) {
 }
 ModalOverlay.propTypes = {
   children: PropTypes.element.isRequired,
+  onClose: PropTypes.func.isRequired,
 };
 export default ModalOverlay;
