+++ src/components/modal/modal.jsx
@@ -4,12 +4,22 @@ import ModalOverlay from "../modal-overlay/modal-overlay";
 
 import { CloseIcon } from "@ya.praktikum/react-developer-burger-ui-components";
 import PropTypes from "prop-types";
-import { useDispatch } from "react-redux";
-// import { handleChangeStatusModal } from "../../services/actions/popup";
+import React, { useEffect } from "react";
+const ESC_KEYCODE = 27;
 const modalRoot = document.getElementById("react-modals");
 function Modal(props) {
   const { children, onClose } = props;
-  const dispatch = useDispatch();
+  useEffect(() => {
+    const closeModal = (e) => {
+      if (e.key === "Escape" || e.keyCode === ESC_KEYCODE) {
+        onClose();
+      }
+    };
+    document.addEventListener("keydown", closeModal);
+    return () => {
+      document.removeEventListener("keydown", closeModal);
+    };
+  }, [onClose]);
 
   return ReactDOM.createPortal(
     <>
@@ -18,10 +28,7 @@ function Modal(props) {
           className={modalSt.modal__inner}
           onClick={(e) => e.stopPropagation(e)}
         >
-          <div
-            className={modalSt.modal__close}
-            onClick={() => dispatch(onClose())}
-          >
+          <div className={modalSt.modal__close} onClick={() => onClose()}>
             <CloseIcon type="primary" />
           </div>
 
@@ -34,5 +41,7 @@ function Modal(props) {
 }
 Modal.propTypes = {
   children: PropTypes.element.isRequired,
+  onClose: PropTypes.func.isRequired,
 };
+
 export default Modal;
